Using prefabs in Forced Choice
Used Motion Tracking and Visual Conflict prefabs that were created in
Grasping in the Forced Choice scene. NB It’s a bit of a pain because
lots of states in the Visual Conflict FSMs had to be reconnected.
Perhaps a single prefab that holds both th e Motion Tracking and Visual
Conflict game objects would maintain the connections.